optimize DateUtils
