--- src/main/java/com/scrats/rent/util/DateUtil.java
@@ -1,234 +1,0 @@
-package com.scrats.rent.util;
-
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
-
-/**
- * 说明：日期处理
- * 创建人：FH Q313596790
- * 修改时间：2015年11月24日
- */
-public class DateUtil {
-
-    private final static SimpleDateFormat sdfYear = new SimpleDateFormat("yyyy");
-    private final static SimpleDateFormat sdfMonth = new SimpleDateFormat("yyyyMM");
-    private final static SimpleDateFormat sdfDay = new SimpleDateFormat("yyyy-MM-dd");
-    private final static SimpleDateFormat sdfDays = new SimpleDateFormat("yyyyMMdd");
-    private final static SimpleDateFormat sdfTime = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-    private final static SimpleDateFormat sdfTimes = new SimpleDateFormat("yyyyMMddHHmmss");
-
-    /**
-     * 获取YYYY格式
-     */
-    public static String getYear(Date date) {
-
-        if(null == date)
-            date = new Date();
-        return sdfYear.format(date);
-    }
-
-    /**
-     * 获取YYYYMM格式
-     */
-    public static String getMonth(Date date) {
-        if(null == date)
-            date = new Date();
-        return sdfMonth.format(date);
-    }
-
-    /**
-     * 获取YYYY-MM-DD格式
-     */
-    public static String getDay(Date date) {
-        if(null == date)
-            date = new Date();
-        return sdfDay.format(date);
-    }
-
-    /**
-     * 获取YYYYMMDD格式
-     */
-    public static String getDays(Date date) {
-        if(null == date)
-            date = new Date();
-        return sdfDays.format(date);
-    }
-
-    /**
-     * 获取YYYY-MM-DD HH:mm:ss格式
-     */
-    public static String getTime(Date date) {
-        if(null == date)
-            date = new Date();
-        return sdfTime.format(date);
-    }
-
-    /**
-     * 获取YYYY格式
-     */
-    public static String getTimes(Date date) {
-        if(null == date)
-            date = new Date();
-        return sdfTimes.format(date);
-    }
-
-    /**
-     * @return boolean
-     * @throws
-     * @Title: compareDate
-     * @Description: TODO(日期比较，如果s>=e 返回true 否则返回false)
-     * @author fh
-     */
-    public static boolean compareDate(String s, String e) {
-        if (fomatDate(s) == null || fomatDate(e) == null) {
-            return false;
-        }
-        return fomatDate(s).getTime() >= fomatDate(e).getTime();
-    }
-
-    /**
-     * 格式化日期
-     */
-    public static Date fomatDate(String date) {
-        DateFormat fmt = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            return fmt.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
-     * 格式化日期
-     */
-    public static Date fomatDate2(String date) {
-        try {
-            return sdfTime.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
-     * 校验日期是否合法
-     */
-    public static boolean isValidDate(String s) {
-        DateFormat fmt = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            fmt.parse(s);
-            return true;
-        } catch (Exception e) {
-            // 如果throw java.text.ParseException或者NullPointerException，就说明格式不对
-            return false;
-        }
-    }
-
-    /**
-     * @param startTime
-     * @param endTime
-     * @return
-     */
-    public static int getDiffYear(String startTime, String endTime) {
-        DateFormat fmt = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            //long aa=0;
-            int years = (int) (((fmt.parse(endTime).getTime() - fmt.parse(startTime).getTime()) / (1000
-                    * 60
-                    * 60
-                    * 24)) / 365);
-            return years;
-        } catch (Exception e) {
-            // 如果throw java.text.ParseException或者NullPointerException，就说明格式不对
-            return 0;
-        }
-    }
-
-    /**
-     * <li>功能描述：时间相减得到天数
-     *
-     * @return long
-     * @author Administrator
-     */
-    public static long getDaySub(String beginDateStr, String endDateStr) {
-        long day = 0;
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-        Date beginDate = null;
-        Date endDate = null;
-
-        try {
-            beginDate = format.parse(beginDateStr);
-            endDate = format.parse(endDateStr);
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        day = (endDate.getTime() - beginDate.getTime()) / (24 * 60 * 60 * 1000);
-        //System.out.println("相隔的天数="+day);
-
-        return day;
-    }
-
-    /**
-     * 得到n天之后的日期
-     */
-    public static String getAfterDayDate(String days) {
-        int daysInt = Integer.parseInt(days);
-
-        Calendar canlendar = Calendar.getInstance(); // java.util包
-        canlendar.add(Calendar.DATE, daysInt); // 日期减 如果不够减会将月变动
-        Date date = canlendar.getTime();
-
-        SimpleDateFormat sdfd = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        String dateStr = sdfd.format(date);
-
-        return dateStr;
-    }
-
-    /**
-     * 得到n天之后是周几
-     */
-    public static String getAfterDayWeek(String days) {
-        int daysInt = Integer.parseInt(days);
-        Calendar canlendar = Calendar.getInstance(); // java.util包
-        canlendar.add(Calendar.DATE, daysInt); // 日期减 如果不够减会将月变动
-        Date date = canlendar.getTime();
-        SimpleDateFormat sdf = new SimpleDateFormat("E");
-        String dateStr = sdf.format(date);
-        return dateStr;
-    }
-
-    public static String getYearOfBeforeMonth(int i) {
-        Calendar canlendar = Calendar.getInstance();
-        canlendar.add(Calendar.MONTH, i);
-        return canlendar.get(Calendar.YEAR) + "";
-    }
-
-    public static String getMonthOfBeforeMonth(int i) {
-        Calendar canlendar = Calendar.getInstance();
-        canlendar.add(Calendar.MONTH, i);
-        return canlendar.get(Calendar.MONTH) + 1 + "";
-    }
-
-    public static String getBeforeMonth(int i) {
-        Calendar canlendar = Calendar.getInstance();
-        canlendar.add(Calendar.MONTH, i);
-        return sdfMonth.format(canlendar.getTime());
-    }
-
-    /**
-     * 格式化时分秒
-     */
-    public static Date string_to_HMSDate(String date) {
-        DateFormat fmt = new SimpleDateFormat("HH:mm:ss");
-        try {
-            return fmt.parse(date);
-        } catch (ParseException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-}
